Refactor API helper methods.
Break up helpers.go into specific files and make some style more
consistent.
